requirements.txt
# Taken from Monai

# Base Monai requirements
torch>=1.9
numpy>=1.24,<2.0

# "min" Monai requirements
setuptools>=50.3.0,<66.0.0,!=60.6.0 ; python_version < "3.12"
setuptools>=70.2.0; python_version >= "3.12"
coverage>=5.5
parameterized
packaging

# "dev" Monai requirements
pytorch-ignite==0.4.11
gdown>=4.7.3
scipy>=1.12.0; python_version >= '3.9'
itk>=5.2
nibabel
pillow!=8.3.0  # https://github.com/python-pillow/Pillow/issues/5571
tensorboard>=2.12.0  # https://github.com/Project-MONAI/MONAI/issues/7434
scikit-image>=0.19.0
tqdm>=4.47.0
lmdb
flake8>=3.8.1
flake8-bugbear<=24.2.6  # https://github.com/Project-MONAI/MONAI/issues/7690
flake8-comprehensions
mccabe
pep8-naming
pycodestyle
pyflakes
black>=22.12
isort>=5.1
ruff
pytype>=2020.6.1; platform_system != "Windows"
types-setuptools
mypy>=1.5.0, <1.12.0
ninja
torchio
torchvision
psutil
cucim-cu12; platform_system == "Linux" and python_version >= "3.9" and python_version <= "3.10"
openslide-python
imagecodecs; platform_system == "Linux" or platform_system == "Darwin"
tifffile; platform_system == "Linux" or platform_system == "Darwin"
pandas
requests
einops
transformers>=4.36.0, <4.41.0; python_version <= '3.10'
mlflow>=2.12.2
clearml>=1.10.0rc0
matplotlib>=3.6.3
tensorboardX
types-PyYAML
pyyaml
fire
jsonschema
pynrrd
pre-commit
pydicom
h5py
nni==2.10.1; platform_system == "Linux" and "arm" not in platform_machine and "aarch" not in platform_machine
optuna
git+https://github.com/Project-MONAI/MetricsReloaded@monai-support#egg=MetricsReloaded
onnx>=1.13.0
onnxruntime; python_version <= '3.10'
typeguard<3  # https://github.com/microsoft/nni/issues/5457
filelock<3.12.0  # https://github.com/microsoft/nni/issues/5523
zarr
lpips==0.1.4
nvidia-ml-py
huggingface_hub
pyamg>=5.0.0
git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
onnx_graphsurgeon
polygraphy
